requirements.txt
discord.py>=2.3.0
aiosqlite>=0.18.0
